+++ pyvistaqt/__init__.py
@@ -1,3 +1,6 @@
 """PyVista package for 3D plotting and mesh analysis."""
 from pyvistaqt._version import __version__
-from pyvistaqt.plotting import BackgroundPlotter, MainWindow, QtInteractor
+from pyvistaqt.plotting import BackgroundPlotter, QtInteractor
+from pyvistaqt.window import MainWindow
+from pyvistaqt.counter import Counter
+from pyvistaqt.dialog import FileDialog, ScaleAxesDialog
